add title to list visual
