+++ packages/lesmis/src/components/Visuals/index.js
@@ -1,0 +1,7 @@
+/*
+ * Tupaia
+ *  Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ */
+
+export * from './Chart';
+export * from './ListVisual';
